+++ Mage/Map/Overlay/LocationAccuracyRenderer.h
@@ -1,0 +1,18 @@
+//
+//  LcoationAccuracyRenderer.h
+//  MAGE
+//
+//  Created by William Newman on 5/6/20.
+//  Copyright © 2020 National Geospatial Intelligence Agency. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <MapKit/MapKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface LocationAccuracyRenderer : MKCircleRenderer
+
+@end
+
+NS_ASSUME_NONNULL_END