+++ Mage/Map/Overlay/ObservationAccuracy.m
@@ -1,0 +1,13 @@
+//
+//  AccuracyOverlay.m
+//  MAGE
+//
+//  Created by William Newman on 5/6/20.
+//  Copyright © 2020 National Geospatial Intelligence Agency. All rights reserved.
+//
+
+#import "ObservationAccuracy.h"
+
+@implementation ObservationAccuracy
+
+@end